fix(leaks,locks,build): 누수/락 해결 및 전 빌드 warning 제거
누수:
- host_shutdown_internal()에서 nickname_reserve_lock,
  version_ip_ban_rules, reserved_nicknames 정리 누락 수정
- nickname_reserve_lock_initialized 플래그 추가

Lock state:
- session_channel_write_all() SSH 경로에서 channel NULL 재확인
  (session_close_channel()과의 TOCTOU 완화)
- session_destroy()의 malloc_trim(0) 제거
  (다수 세션 종료 시 글로벌 lock 경합 방지)

빌드/도구:
- Makefile에 debug, asan, tsan, valgrind 타겟 추가
  (gdb, kcachegrind, AddressSanitizer, ThreadSanitizer 즉시 사용 가능)
- _FORTIFY_SOURCE=3 → 2 (Slackware 15.0 glibc 호환)
- --warn-execstack 제거 (GNU ld 2.37 미지원)
- -lutil 추가 (forkpty 링크)
- src/host.c에 _GNU_SOURCE 추가

Warning 제거:
- format-truncation: host.c 단위 pragma + morse_client.c 길이 제한
- format-nonliteral: translator.c pragma push/pop
- sign-conversion: host_session_output_messaging.c 명시적 캐스팅

검증: make / make debug / make valgrind / make asan
모두 Slackware 15.0에서 warning 0, 링크 성공

+++ src/host_parts/output/host_session_output_messaging.c
@@ -275,17 +275,21 @@ int count_unicode_points(const char *str, utf8_code_count_t **counts_out,
             bytes_read = 1;
         } else if ((c & 0xE0) == 0xC0) {
             if (i + 1 < len)
-                code_point = ((c & 0x1F) << 6) | (str[i + 1] & 0x3F);
+                code_point = ((unsigned int)(c & 0x1F) << 6) |
+                             (unsigned int)(str[i + 1] & 0x3F);
             bytes_read = 2;
         } else if ((c & 0xF0) == 0xE0) {
             if (i + 2 < len)
-                code_point = ((c & 0x0F) << 12) | ((str[i + 1] & 0x3F) << 6) |
-                             (str[i + 2] & 0x3F);
+                code_point = ((unsigned int)(c & 0x0F) << 12) |
+                             ((unsigned int)(str[i + 1] & 0x3F) << 6) |
+                             (unsigned int)(str[i + 2] & 0x3F);
             bytes_read = 3;
         } else if ((c & 0xF8) == 0xF0) {
             if (i + 3 < len)
-                code_point = ((c & 0x07) << 18) | ((str[i + 1] & 0x3F) << 12) |
-                             ((str[i + 2] & 0x3F) << 6) | (str[i + 3] & 0x3F);
+                code_point = ((unsigned int)(c & 0x07) << 18) |
+                             ((unsigned int)(str[i + 1] & 0x3F) << 12) |
+                             ((unsigned int)(str[i + 2] & 0x3F) << 6) |
+                             (unsigned int)(str[i + 3] & 0x3F);
             bytes_read = 4;
         } else {
             code_point = c;
